# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_osi-oss_osi_internal_dto.CreateLocationRequest:
    description: Данные для создания новой локации (офиса, филиала, склада и т.д.)
    properties:
      address:
        description: |-
          Физический адрес локации
          Example: г. Москва, ул. Ленина, д. 10
        example: г. Москва, ул. Ленина, д. 10
        type: string
      is_verified:
        description: |-
          Подтверждена ли локация
          Example: false
        example: false
        type: boolean
      name:
        description: |-
          Название локации (обязательное)
          Example: Главный офис
        example: Главный офис
        type: string
      source:
        description: |-
          Источник данных: manual, egrul, api
          Example: manual
        enum:
        - manual
        - egrul
        - api
        example: manual
        type: string
    required:
    - name
    - source
    type: object
...
  github_com_osi-oss_osi_internal_dto.ErrorResponse:
    description: Стандартный формат ответа при ошибке
    properties:
      error:
        description: |-
          Сообщение об ошибке
          Example: invalid email or password
        example: invalid email or password
        type: string
    type: object
  github_com_osi-oss_osi_internal_dto.LocationResponse:
    description: Полная информация о локации
    properties:
      address:
        description: |-
          Физический адрес
          Example: г. Москва, ул. Ленина, д. 10
        example: г. Москва, ул. Ленина, д. 10
        type: string
      created_at:
        description: |-
          Дата создания
          Example: 2024-01-15T10:30:00Z
        example: "2024-01-15T10:30:00Z"
        type: string
      id:
        description: |-
          Уникальный идентификатор локации
          Example: 1
        example: 1
        type: integer
      is_active:
        description: |-
          Активна ли локация
          Example: true
        example: true
        type: boolean
      is_verified:
        description: |-
          Подтверждена ли локация
          Example: true
        example: true
        type: boolean
      name:
        description: |-
          Название локации
          Example: Главный офис
        example: Главный офис
        type: string
      organization_id:
        description: |-
          ID организации, которой принадлежит локация
          Example: 1
        example: 1
        type: integer
      source:
        description: |-
          Источник данных
          Example: manual
        enum:
        - manual
        - egrul
        - api
        example: manual
        type: string
      updated_at:
        description: |-
          Дата последнего обновления
          Example: 2024-01-20T15:45:00Z
        example: "2024-01-20T15:45:00Z"
        type: string
    type: object
  github_com_osi-oss_osi_internal_dto.LocationsListResponse:
    description: Список локаций организации
    properties:
      locations:
        description: Массив локаций
        items:
          $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.LocationResponse'
        type: array
    type: object
...
  github_com_osi-oss_osi_internal_dto.MessageResponse:
    description: Стандартный формат успешного ответа с сообщением
    properties:
      message:
        description: |-
          Сообщение о результате операции
          Example: operation completed successfully
        example: operation completed successfully
        type: string
    type: object
...
  github_com_osi-oss_osi_internal_dto.UpdateLocationRequest:
    description: Данные для обновления локации
    properties:
      address:
        description: |-
          Новый адрес
          Example: г. Москва, ул. Пушкина, д. 5
        example: г. Москва, ул. Пушкина, д. 5
        type: string
      is_active:
        description: |-
          Активна ли локация
          Example: true
        example: true
        type: boolean
      is_verified:
        description: |-
          Статус верификации
          Example: true
        example: true
        type: boolean
      name:
        description: |-
          Новое название локации
          Example: Центральный офис
        example: Центральный офис
        type: string
      source:
        description: |-
          Источник данных
          Example: manual
        enum:
        - manual
        - egrul
        - api
        example: manual
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для управления организациями, локациями, отделами и позициями
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: OSI API
  version: "1.0"
paths:
...
  /organizations/{orgId}/locations:
    get:
      consumes:
      - application/json
      description: |-
        Возвращает все локации организации (активные и неактивные).
        Требуется доступ к организации (участник или основатель).
      parameters:
      - description: ID организации
        example: 1
        in: path
        minimum: 1
        name: orgId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список локаций организации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.LocationsListResponse'
        "400":
          description: Неверный формат ID организации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "401":
          description: Отсутствует или невалидный токен авторизации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "403":
          description: Нет доступа к организации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список локаций организации
      tags:
      - Локации
    post:
      consumes:
      - application/json
      description: |-
        Создаёт новую локацию (офис, филиал, склад и т.д.) в организации.
        Требуется право "locations.create" или статус основателя.
        Поле source указывает источник данных: manual (вручную), egrul (из ЕГРЮЛ), api (через API).
      parameters:
      - description: ID организации
        example: 1
        in: path
        minimum: 1
        name: orgId
        required: true
        type: integer
      - description: Данные для создания локации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.CreateLocationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Локация успешно создана
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.LocationResponse'
        "400":
          description: 'Ошибка валидации: name и source обязательны'
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "401":
          description: Отсутствует или невалидный токен авторизации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "403":
          description: Нет права locations.create
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "404":
          description: Организация не найдена
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создание новой локации
      tags:
      - Локации
  /organizations/{orgId}/locations/{locId}:
    delete:
      consumes:
      - application/json
      description: |-
        Удаляет локацию из организации.
        **ВНИМАНИЕ**: Удаление локации удалит все связанные отделы.
        Требуется право "locations.delete" или статус основателя.
      parameters:
      - description: ID организации
        example: 1
        in: path
        minimum: 1
        name: orgId
        required: true
        type: integer
      - description: ID локации
        example: 1
        in: path
        minimum: 1
        name: locId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Локация успешно удалена
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.MessageResponse'
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "401":
          description: Отсутствует или невалидный токен авторизации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "403":
          description: Нет права locations.delete
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "404":
          description: Локация не найдена
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление локации
      tags:
      - Локации
    get:
      consumes:
      - application/json
      description: |-
        Возвращает полную информацию о локации.
        Требуется доступ к организации.
      parameters:
      - description: ID организации
        example: 1
        in: path
        minimum: 1
        name: orgId
        required: true
        type: integer
      - description: ID локации
        example: 1
        in: path
        minimum: 1
        name: locId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные локации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.LocationResponse'
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "401":
          description: Отсутствует или невалидный токен авторизации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "404":
          description: Локация не найдена
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение локации по ID
      tags:
      - Локации
    put:
      consumes:
      - application/json
      description: |-
        Обновляет данные локации.
        Можно обновить название, адрес, источник, статус верификации и активность.
        Требуется право "locations.update" или статус основателя.
      parameters:
      - description: ID организации
        example: 1
        in: path
        minimum: 1
        name: orgId
        required: true
        type: integer
      - description: ID локации
        example: 1
        in: path
        minimum: 1
        name: locId
        required: true
        type: integer
      - description: Новые данные локации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.UpdateLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Локация успешно обновлена
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.LocationResponse'
        "400":
          description: Ошибка валидации данных
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "401":
          description: Отсутствует или невалидный токен авторизации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "403":
          description: Нет права locations.update
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "404":
          description: Локация не найдена
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление локации
      tags:
      - Локации
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"